[FE][JL] Restructure folder based on new conventions
