+++ src/components/fields/textarea/textarea.tsx
@@ -1,8 +1,8 @@
 import { kebabCase } from "lodash";
 import React, { useEffect, useRef } from "react";
 import { Form } from "react-lifesg-design-system";
-import { IGenericFieldProps } from "../../types";
-import { InteractionHelper, TestHelper } from "../../utils";
+import { InteractionHelper, TestHelper } from "../../../helpers";
+import { IGenericFieldProps } from "../../frontend-engine/types";
 import { AutoResizeTextarea } from "./auto-resize-textarea";
 import { ChipContainer, ChipItem } from "./textarea.styles";
 import { ITextareaProps } from "./types";
