+++ src/components/frontend-engine/index.ts
@@ -1,2 +1,2 @@
-export * from "./fields";
 export * from "./frontend-engine";
+export * from "./types";
