+++ src/helpers/interaction-helper.ts
