+++ src/index.ts
@@ -1,2 +1,1 @@
-export * from "./frontend-engine";
-export * from "./types";
+export * from "./components";
